[Refactor] album 패키지 생성 > > Co-authored-by: Yujin Jeong <yujin45@sookmyung.ac.kr> Co-authored-by: lh99j <grand2181@gmail.com>
